renaming SimpleTest classes to Default
