+++ Benchmarks/ExampleSetups/FixedSalinityFromFile/FixedSalinity_FromFile_C.xml
@@ -2,7 +2,7 @@
     <random_seed>643879</random_seed>
     <tree_dynamics>
         <aboveground_competition>
-            <type> SimpleTest </type>
+            <type> Default </type>
         </aboveground_competition>
         <belowground_competition>
             <type> FixedSalinity </type>
