+++ Benchmarks/ModuleBenchmarks/PlantModules/SimpleBettina/SimpleTest/SimpleTest_CI.xml
@@ -2,10 +2,10 @@
     <random_seed>643879</random_seed>
     <resources>
         <aboveground>
-            <type> SimpleTest </type>
+            <type> Default </type>
         </aboveground>
         <belowground>
-            <type> SimpleTest </type>
+            <type> Default </type>
         </belowground>
     </resources>
     <plant_dynamics>
