+++ ResourceLib/AboveGround/Default/Default.py
@@ -2,13 +2,13 @@
 # -*- coding: utf-8 -*-
 """
 @date: 2018-Today
-@author: jasper.bathmann@ufz.de
+@author: jbathmann
 """
 import numpy as np
 from ResourceLib import ResourceModel
 
 
-class SimpleTest(ResourceModel):
+class Default(ResourceModel):
 
     def __init__(self, args):
         ## SimpleTest case for aboveground competition concept. This case is
