+++ ResourceLib/AboveGround/__init__.py
@@ -6,5 +6,5 @@ Created on Thu Nov  8 15:25:03 2018
 @author: bathmann
 """
 
-from .SimpleTest import SimpleTest
+from .Default import Default
 from .SimpleAsymmetricZOI import SimpleAsymmetricZOI
